Split out generation functions into their own handler and helper files
