+++ dataactbroker/handlers/fileHandler.py
@@ -33,8 +33,8 @@ from dataactcore.models.domainModels import CGAC, FREC, SubTierAgency, States, C
 from dataactcore.models.jobModels import (Job, Submission, SubmissionNarrative, SubmissionSubTierAffiliation,
                                           RevalidationThreshold, CertifyHistory, CertifiedFilesHistory, FileRequest)
 from dataactcore.models.lookups import (
-    FILE_TYPE_DICT, FILE_TYPE_DICT_LETTER, FILE_TYPE_DICT_LETTER_ID, PUBLISH_STATUS_DICT, JOB_TYPE_DICT,
-    FILE_TYPE_DICT_ID, JOB_STATUS_DICT, JOB_STATUS_DICT_ID, PUBLISH_STATUS_DICT_ID, FILE_TYPE_DICT_LETTER_NAME)
+    FILE_TYPE_DICT_LETTER, FILE_TYPE_DICT_LETTER_ID, PUBLISH_STATUS_DICT, JOB_TYPE_DICT, FILE_TYPE_DICT,
+    JOB_STATUS_DICT, JOB_STATUS_DICT_ID, PUBLISH_STATUS_DICT_ID, FILE_TYPE_DICT_LETTER_NAME)
 from dataactcore.models.stagingModels import (DetachedAwardFinancialAssistance, PublishedAwardFinancialAssistance,
                                               FPDSContractingOffice)
 from dataactcore.models.userModel import User
@@ -49,8 +49,6 @@ from dataactcore.utils.statusCode import StatusCode
 from dataactcore.utils.stringCleaner import StringCleaner
 
 from dataactvalidator.filestreaming.csv_selection import write_query_to_file
-from dataactvalidator.validation_handlers.file_generation_handler import (
-    check_file_generation, start_d_generation, start_e_f_generation)
 
 logger = logging.getLogger(__name__)
 
@@ -450,136 +448,6 @@ class FileHandler:
             copyfile(file_url, local_file_path)
             return True
 
-    def generate_file(self, submission, file_type, start, end, agency_type):
-        """ Start a file generation job for the specified file type within a submission
-
-            Args:
-                submission: submission for which we're generating the file
-                file_type: type of file to generate the job for
-                start: the start date for the file to generate
-                end: the end date for the file to generate
-                agency_type: The type of agency (awarding or funding) to generate the file for (only used for D file
-                    generation)
-
-            Returns:
-                Results of check_generation or JsonResponse object containing an error if the prerequisite job isn't
-                complete.
-        """
-        error_message = None
-        # submission is a FABS submission
-        if submission.d2_submission:
-            error_message = "Cannot generate files for FABS submissions."
-
-        elif file_type in ['D1', 'D2']:
-            # D file generation requires start and end date
-            if not start or not end:
-                error_message = "Must have a start and end date for D file generation."
-            # D files can only be generated by awarding or funding agency
-            elif agency_type not in ['awarding', 'funding']:
-                error_message = "agency_type must be either awarding or funding for D file generation."
-
-        # Only D1, D2, E, and F files can be generated
-        elif file_type not in ['E', 'F']:
-            error_message = "File type must be either D1, D2, E, or F"
-
-        # Return any client errors
-        if error_message:
-            return JsonResponse.error(ValueError(error_message), StatusCode.CLIENT_ERROR)
-
-        sess = GlobalDB.db().session
-        job = sess.query(Job).filter(Job.submission_id == submission.submission_id,
-                                     Job.file_type_id == FILE_TYPE_DICT_LETTER_ID[file_type],
-                                     Job.job_type_id == JOB_TYPE_DICT['file_upload']).one()
-
-        log_data = {
-            'message': 'Starting {} file generation within submission {}'.format(file_type, submission.submission_id),
-            'message_type': 'BrokerInfo',
-            'submission_id': submission.submission_id,
-            'job_id': job.job_id,
-            'file_type': file_type
-        }
-        logger.info(log_data)
-
-        # Check prerequisites on upload job
-        if not check_generation_prereqs(submission.submission_id, file_type):
-            return JsonResponse.error(ResponseException("Must wait for completion of prerequisite validation job",
-                                                        StatusCode.CLIENT_ERROR), StatusCode.CLIENT_ERROR)
-        try:
-            if file_type in ['D1', 'D2']:
-                response = start_d_generation(job, start, end, agency_type)
-            else:
-                response = start_e_f_generation(job)
-
-            log_data['message'] = 'SQS message response: {}'.format(response)
-            logger.debug(log_data)
-        except Exception as e:
-            job.job_status_id = JOB_STATUS_DICT['failed']
-            job.error_message = str(e)
-            sess.commit()
-            return JsonResponse.error(e, StatusCode.INTERNAL_ERROR)
-
-        # Return same response as check generation route
-        return self.check_generation(submission, file_type)
-
-    def generate_detached_file(self, file_type, cgac_code, frec_code, start, end, agency_type):
-        """ Start a file generation job for the specified file type not connected to a submission
-
-            Args:
-                file_type: type of file to be generated
-                cgac_code: the code of a CGAC agency if generating for a CGAC agency
-                frec_code: the code of a FREC agency if generating for a FREC agency
-                start: start date in a string, formatted MM/DD/YYYY
-                end: end date in a string, formatted MM/DD/YYYY
-                agency_type: The type of agency (awarding or funding) to generate the file for
-
-            Returns:
-                JSONResponse object with keys job_id, status, file_type, url, message, start, and end.
-
-            Raises:
-                ResponseException: if the start and end Strings cannot be parsed into dates
-        """
-        # Make sure it's a valid request
-        if not cgac_code and not frec_code:
-            return JsonResponse.error(ValueError("Detached file generation requires CGAC or FR Entity Code"),
-                                      StatusCode.CLIENT_ERROR)
-
-        # Check if date format is MM/DD/YYYY
-        if not (StringCleaner.is_date(start) and StringCleaner.is_date(end)):
-            raise ResponseException('Start or end date cannot be parsed into a date', StatusCode.CLIENT_ERROR)
-
-        if agency_type not in ('awarding', 'funding'):
-            return JsonResponse.error(ValueError("agency_type must be either awarding or funding."),
-                                      StatusCode.CLIENT_ERROR)
-
-        # Add job info
-        file_type_name = FILE_TYPE_DICT_ID[FILE_TYPE_DICT_LETTER_ID[file_type]]
-        new_job = self.add_generation_job_info(file_type_name=file_type_name, start_date=start, end_date=end)
-
-        agency_code = frec_code if frec_code else cgac_code
-        log_data = {
-            'message': 'Starting detached {} file generation'.format(file_type),
-            'message_type': 'BrokerInfo',
-            'job_id': new_job.job_id,
-            'file_type': file_type,
-            'agency_code': agency_code,
-            'start_date': start,
-            'end_date': end
-        }
-        logger.info(log_data)
-
-        try:
-            response = start_d_generation(new_job, start, end, agency_type, agency_code=agency_code)
-            log_data['message'] = 'SQS message response: {}'.format(response)
-            logger.debug(log_data)
-        except Exception as e:
-            new_job.job_status_id = JOB_STATUS_DICT['failed']
-            new_job.error_message = str(e)
-            GlobalDB.db().session.commit()
-            return JsonResponse.error(e, StatusCode.INTERNAL_ERROR)
-
-        # Return same response as check generation route
-        return self.check_detached_generation(new_job.job_id)
-
     def upload_fabs_file(self, fabs):
         """ Uploads the provided FABS file to S3 and creates a new submission if one doesn't exist or updates the
             existing submission if one does.
@@ -691,41 +559,6 @@ class FileHandler:
                 sess.commit()
 
             return json_response
-
-    @staticmethod
-    def check_detached_generation(job_id):
-        """ Return information about detached file generation jobs
-
-            Args:
-                job_id: ID of the detached generation job
-
-            Returns:
-                Response object with keys job_id, status, file_type, url, message, start, and end.
-        """
-        response_dict = check_file_generation(job_id)
-
-        return JsonResponse.create(StatusCode.OK, response_dict)
-
-    @staticmethod
-    def check_generation(submission, file_type):
-        """ Return information about file generation jobs connected to a submission
-
-            Args:
-                submission: submission to get information from
-                file_type: type of file being generated to check on
-
-            Returns:
-                Response object with keys status, file_type, url, message.
-                If file_type is D1 or D2, also includes start and end.
-        """
-        sess = GlobalDB.db().session
-        upload_job = sess.query(Job).filter(Job.submission_id == submission.submission_id,
-                                            Job.file_type_id == FILE_TYPE_DICT_LETTER_ID[file_type],
-                                            Job.job_type_id == JOB_TYPE_DICT['file_upload']).one()
-
-        response_dict = check_file_generation(upload_job.job_id)
-
-        return JsonResponse.create(StatusCode.OK, response_dict)
 
     @staticmethod
     def publish_fabs_submission(submission):
