+++ dataactbroker/routes/generation_routes.py
@@ -1,9 +1,8 @@
-from flask import request
 from webargs import fields as webargs_fields, validate as webargs_validate
 from webargs.flaskparser import use_kwargs
 
 from dataactbroker.decorators import convert_to_submission_id
-from dataactbroker.handlers.fileHandler import FileHandler
+from dataactbroker.handlers import generation_handler
 from dataactbroker.permissions import requires_login, requires_submission_perms
 
 DATE_REGEX = '^\d{2}\/\d{2}\/\d{4}$'
@@ -45,8 +44,7 @@ def add_generation_routes(app, is_local, server_path):
                 end: the end date for the file to generate
                 agency_type: The type of agency (awarding or funding) to generate the file for
         """
-        file_manager = FileHandler(request, is_local=is_local, server_path=server_path)
-        return file_manager.generate_file(submission_id, file_type, start, end, agency_type)
+        return generation_handler.generate_file(submission_id, file_type, start, end, agency_type)
 
     @app.route("/v1/check_generation_status/", methods=["GET"])
     @convert_to_submission_id
@@ -62,8 +60,7 @@ def add_generation_routes(app, is_local, server_path):
                 submission: submission for which we're generating the file
                 file_type: type of file to check the status of
         """
-        file_manager = FileHandler(request, is_local=is_local, server_path=server_path)
-        return file_manager.check_generation(submission, file_type)
+        return generation_handler.check_generation(submission, file_type)
 
     @app.route("/v1/generate_detached_file/", methods=["POST"])
     @requires_login
@@ -90,13 +87,11 @@ def add_generation_routes(app, is_local, server_path):
                 end: end date in a string, formatted MM/DD/YYYY
                 agency_type: The type of agency (awarding or funding) to generate the file for
         """
-        file_manager = FileHandler(request, is_local=is_local, server_path=server_path)
-        return file_manager.generate_detached_file(file_type, cgac_code, frec_code, start, end, agency_type)
+        return generation_handler.generate_detached_file(file_type, cgac_code, frec_code, start, end, agency_type)
 
     @app.route("/v1/check_detached_generation_status/", methods=["GET"])
     @requires_login
     @use_kwargs({'job_id': webargs_fields.Int(required=True)})
     def check_detached_generation_status(job_id):
         """ Return status of file generation job """
-        file_manager = FileHandler(request, is_local=is_local, server_path=server_path)
-        return file_manager.check_detached_generation(job_id)
+        return generation_handler.check_detached_generation(job_id)
