+++ dataactvalidator/validation_handlers/file_generation_handler.py
@@ -7,19 +7,14 @@ from datetime import datetime
 from flask import Flask
 
 from dataactcore.aws.s3Handler import S3Handler
-from dataactcore.aws.sqsHandler import sqs_queue
 from dataactcore.config import CONFIG_BROKER
 from dataactcore.interfaces.db import GlobalDB
 from dataactcore.interfaces.function_bag import mark_job_status
 from dataactcore.models.domainModels import ExecutiveCompensation
-from dataactcore.models.jobModels import Job, FileRequest, FPDSUpdate, Submission
-from dataactcore.models.lookups import (
-    PUBLISH_STATUS_DICT, JOB_STATUS_DICT, JOB_STATUS_DICT_ID, JOB_TYPE_DICT, FILE_TYPE_DICT_LETTER)
+from dataactcore.models.jobModels import Job, FileRequest, FPDSUpdate
+from dataactcore.models.lookups import JOB_STATUS_DICT, JOB_STATUS_DICT_ID, JOB_TYPE_DICT
 from dataactcore.models.stagingModels import AwardFinancialAssistance, AwardProcurement
 from dataactcore.utils import fileD1, fileD2, fileE, fileF
-from dataactcore.utils.responseException import ResponseException
-from dataactcore.utils.statusCode import StatusCode
-from dataactcore.utils.stringCleaner import StringCleaner
 
 from dataactvalidator.filestreaming.csv_selection import write_csv, write_query_to_file, stream_file_to_s3
 
@@ -391,191 +386,3 @@ def copy_file_from_parent_to_child(child_job, parent_job, is_local):
         logger.info(log_data)
         with smart_open.smart_open(S3Handler.create_file_path(parent_job.filename), 'r') as reader:
             stream_file_to_s3(child_job.filename, reader)
-
-
-def start_d_generation(job, start_date, end_date, agency_type, agency_code=None):
-    """ Validates the start and end dates of the generation, updates the submission's publish status and progress (if
-        its not detached generation), and sends the job information to SQS.
-
-        Args:
-            job: File generation job to start
-            start_date: String to parse as the start date of the generation
-            end_date: String to parse as the end date of the generation
-            agency_type: Type of Agency to generate files by: "awarding" or "funding"
-            agency_code: Agency code for detached D file generations
-
-        Returns:
-            SQS send_message response
-    """
-    if not (StringCleaner.is_date(start_date) and StringCleaner.is_date(end_date)):
-        raise ResponseException("Start or end date cannot be parsed into a date of format MM/DD/YYYY",
-                                StatusCode.CLIENT_ERROR)
-
-    # Update the Job's start and end dates
-    sess = GlobalDB.db().session
-    job.start_date = start_date
-    job.end_date = end_date
-    sess.commit()
-
-    # Update submission
-    if job.submission_id:
-        submission = sess.query(Submission).filter(Submission.submission_id == job.submission_id).one()
-
-        # Change the publish status back to updated if certified
-        if submission.publish_status_id == PUBLISH_STATUS_DICT['published']:
-            submission.publishable = False
-            submission.publish_status_id = PUBLISH_STATUS_DICT['updated']
-            submission.updated_at = datetime.utcnow()
-            sess.commit()
-
-        # Set cross-file validation status to waiting if it's not already
-        cross_file_job = sess.query(Job).filter(Job.submission_id == job.submission_id,
-                                                Job.job_type_id == JOB_TYPE_DICT['validation'],
-                                                Job.job_status_id != JOB_STATUS_DICT['waiting']).one_or_none()
-
-        # No need to update it for each type of D file generation job, just do it once
-        if cross_file_job:
-            cross_file_job.job_status_id = JOB_STATUS_DICT['waiting']
-            sess.commit()
-
-    mark_job_status(job.job_id, "waiting")
-
-    log_data = {
-        'message': 'Sending {} file generation job {} to SQS'.format(job.file_type.letter_name, job.job_id),
-        'message_type': 'BrokerInfo',
-        'submission_id': job.submission_id,
-        'job_id': job.job_id,
-        'file_type': job.file_type.letter_name
-    }
-    logger.info(log_data)
-
-    # Set SQS message attributes
-    message_attr = {'agency_type': {'DataType': 'String', 'StringValue': agency_type}}
-    if agency_code:
-        message_attr['agency_code'] = {'DataType': 'String', 'StringValue': agency_code}
-
-    # Add job_id to the SQS job queue
-    queue = sqs_queue()
-    return queue.send_message(MessageBody=str(job.job_id), MessageAttributes=message_attr)
-
-
-def start_e_f_generation(job):
-    """ Passes the Job ID for an E or F generation Job to SQS
-
-        Args:
-            job: File generation job to start
-    """
-    mark_job_status(job.job_id, "waiting")
-
-    file_type = job.file_type.letter_name
-    log_data = {
-        'message': 'Sending {} file generation job {} to Validator in SQS'.format(file_type, job.job_id),
-        'message_type': 'BrokerInfo',
-        'submission_id': job.submission_id,
-        'job_id': job.job_id,
-        'file_type': file_type
-    }
-    logger.info(log_data)
-
-    # Add job_id to the SQS job queue
-    queue = sqs_queue()
-    return queue.send_message(MessageBody=str(job.job_id), MessageAttributes={})
-
-
-def check_file_generation(job_id):
-    """ Check the status of a file generation
-
-        Args:
-            job_id: upload Job ID
-        Return:
-            Dict with keys: job_id, status, file_type, message, url, start, end
-    """
-    sess = GlobalDB.db().session
-
-    # We want to use one_or_none() here so we can see if the job is None so we can mark the status as invalid to
-    # indicate that a status request is invoked for a job that isn't created yet
-    upload_job = sess.query(Job).filter_by(job_id=job_id).one_or_none()
-    response_dict = {'job_id': job_id, 'status': '', 'file_type': '', 'message': '', 'url': '#', 'size': None}
-
-    if upload_job is None:
-        response_dict['start'] = ''
-        response_dict['end'] = ''
-        response_dict['status'] = 'invalid'
-        response_dict['message'] = 'No generation job found with the specified ID'
-        return response_dict
-
-    response_dict['file_type'] = FILE_TYPE_DICT_LETTER[upload_job.file_type_id]
-    response_dict['size'] = upload_job.file_size
-    response_dict['status'] = map_generate_status(sess, upload_job)
-    response_dict['message'] = upload_job.error_message or ''
-
-    # Generate the URL (or path) to the file
-    if CONFIG_BROKER['use_aws'] and response_dict['status'] is 'finished' and upload_job.filename:
-        path, file_name = upload_job.filename.split('/')
-        response_dict['url'] = S3Handler().get_signed_url(path=path, file_name=file_name, bucket_route=None,
-                                                          method='GET')
-    elif response_dict['status'] is 'finished' and upload_job.filename:
-        response_dict['url'] = upload_job.filename
-
-    # Only D file generations have start and end dates
-    if response_dict['file_type'] in ['D1', 'D2']:
-        response_dict['start'] = upload_job.start_date.strftime("%m/%d/%Y") if upload_job.start_date is not None else ""
-        response_dict['end'] = upload_job.end_date.strftime("%m/%d/%Y") if upload_job.end_date is not None else ""
-
-    return response_dict
-
-
-def map_generate_status(sess, upload_job):
-    """ Maps job status to file generation statuses expected by frontend. Updates the error message of the job, if
-        there is one.
-
-        Args:
-            upload_job: the upload job for this file
-            validation_job: the validation job for this file if applicable
-
-        Returns:
-            The status of the submission based on upload job status and validation job status (where applicable)
-    """
-    if FILE_TYPE_DICT_LETTER[upload_job.file_type_id] in ['D1', 'D2'] and upload_job.submission_id:
-        validation_job = sess.query(Job).filter(Job.submission_id == upload_job.submission_id,
-                                                Job.file_type_id == upload_job.file_type_id,
-                                                Job.job_type_id == JOB_TYPE_DICT['csv_record_validation']).one()
-        validation_status = validation_job.job_status.name
-        if validation_job.number_of_errors > 0:
-            errors_present = True
-        else:
-            errors_present = False
-    else:
-        validation_job = None
-        errors_present = False
-
-    response_status = STATUS_MAP[upload_job.job_status.name]
-    if response_status == "failed" and upload_job.error_message in ['', None]:
-        # Provide an error message if none present
-        upload_job.error_message = "Upload job failed without error message"
-
-    if validation_job is None:
-        # No validation job, so don't need to check it
-        sess.commit()
-        return response_status
-
-    if response_status == "finished":
-        # Check status of validation job if present
-        response_status = VALIDATION_STATUS_MAP[validation_status]
-        if response_status == "finished" and errors_present:
-            # If validation completed with errors, mark as failed
-            response_status = "failed"
-            upload_job.error_message = "Validation completed but row-level errors were found"
-
-    if response_status == "failed":
-        if upload_job.error_message in ['', None] and validation_job.error_message in ['', None]:
-            if validation_status == "invalid":
-                upload_job.error_message = "Generated file had file-level errors"
-            else:
-                upload_job.error_message = "Validation job had an internal error"
-
-        elif upload_job.error_message in ['', None]:
-            upload_job.error_message = validation_job.error_message
-
-    sess.commit()
-    return response_status
